+++ src/lib.rs
@@ -1,31 +1,41 @@
-use std::error::Error;
+//! Go the the [readme](https://crates.io/crates/read_input) file for documentation.
+
+// This lint asks for default when new takes type parameter.
+#![allow(clippy::new_without_default)]
+// `impl ToString` is better than `&impl ToString`. Clippy is not ready for impl trait.
+#![allow(clippy::needless_pass_by_value)]
+
+pub mod prelude;
+pub mod shortcut;
+
 use std::io;
 use std::io::Write;
 use std::str::FromStr;
+use std::string::ToString;
 
 const DEFAULT_ERR: &str = "That value does not pass. Please try again";
 
-struct PromptMsg<'a> {
-    msg: &'a str,
+struct PromptMsg {
+    msg: String,
     repeat: bool,
 }
 
-impl<'a> PromptMsg<'a> {
+impl PromptMsg {
     fn new() -> Self {
         Self {
-            msg: "",
+            msg: String::new(),
             repeat: false,
         }
     }
-    fn from_str(s: &'a str) -> Self {
+    fn from_str(s: impl ToString) -> Self {
         Self {
-            msg: s,
+            msg: s.to_string(),
             repeat: false,
         }
     }
-    fn repeat_from_str(s: &'a str) -> Self {
+    fn repeat_from_str(s: impl ToString) -> Self {
         Self {
-            msg: s,
+            msg: s.to_string(),
             repeat: true,
         }
     }
@@ -34,51 +44,57 @@ impl<'a> PromptMsg<'a> {
 type TestFunc<T> = Fn(&T) -> bool;
 
 /// `InputBuilder` is a 'builder' used to store the settings that are used to fetch input.
-pub struct InputBuilder<'a, T: FromStr> {
-    msg: PromptMsg<'a>,
-    err: &'a str,
-    default: Option<T>,
-    test: Vec<(Box<TestFunc<T>>, Option<&'a str>)>,
+pub struct InputBuilder<T: FromStr> {
+    msg: PromptMsg,
+    err: String,
+    test: Vec<(Box<TestFunc<T>>, Option<String>)>,
     err_match: Box<dyn Fn(&T::Err) -> Option<String>>,
 }
 
-impl<'a, T: FromStr> InputBuilder<'a, T> {
+pub struct InputBuilderOnce<T: FromStr> {
+    builder: InputBuilder<T>,
+    default: Option<T>,
+}
+
+impl<T: FromStr> InputBuilder<T> {
     /// Creates a new instance of `InputBuilder` with default settings.
-    pub fn new() -> InputBuilder<'a, T> {
+    pub fn new() -> Self {
         InputBuilder {
             msg: PromptMsg::new(),
-            err: DEFAULT_ERR,
-            default: None,
+            err: DEFAULT_ERR.to_string(),
             test: Vec::new(),
             err_match: Box::new(|_| None),
         }
     }
     /// Changes or adds a prompt message. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn msg(self, msg: &'a str) -> Self {
+    pub fn msg(self, msg: impl ToString) -> Self {
         InputBuilder {
             msg: PromptMsg::from_str(msg),
             ..self
         }
     }
     /// Changes or adds a prompt message and makes it repeat. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn repeat_msg(self, msg: &'a str) -> Self {
+    pub fn repeat_msg(self, msg: impl ToString) -> Self {
         InputBuilder {
             msg: PromptMsg::repeat_from_str(msg),
             ..self
         }
     }
     /// Changes fallback error message. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn err(self, err: &'a str) -> Self {
-        InputBuilder { err, ..self }
-    }
-    /// Changes or adds a default input value. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn default(self, default: T) -> Self {
+    pub fn err(self, err: impl ToString) -> Self {
         InputBuilder {
-            default: Some(default),
+            err: err.to_string(),
             ..self
         }
     }
-    fn test<F: 'static + Fn(&T) -> bool>(self, test: F, err: Option<&'a str>) -> Self {
+    /// Changes or adds a default input value. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
+    pub fn default(self, default: T) -> InputBuilderOnce<T> {
+        InputBuilderOnce {
+            builder: self,
+            default: Some(default),
+        }
+    }
+    fn test<F: 'static + Fn(&T) -> bool>(self, test: F, err: Option<String>) -> Self {
         InputBuilder {
             test: {
                 let mut x = self.test;
@@ -93,8 +109,8 @@ impl<'a, T: FromStr> InputBuilder<'a, T> {
         self.test(test, None)
     }
     /// Adds a validation check on input with custom error message. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-    pub fn add_err_test<F: 'static + Fn(&T) -> bool>(self, test: F, err: &'a str) -> Self {
-        self.test(test, Some(err))
+    pub fn add_err_test<F: 'static + Fn(&T) -> bool>(self, test: F, err: impl ToString) -> Self {
+        self.test(test, Some(err.to_string()))
     }
     /// Removes all validation checks made by `.add_test()` and `.add_err_test()`.
     pub fn clear_tests(self) -> Self {
@@ -111,56 +127,100 @@ impl<'a, T: FromStr> InputBuilder<'a, T> {
         }
     }
     /// 'gets' the input form the user. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
+    pub fn get(&self) -> T {
+        read_input::<T>(&self.msg, &self.err, None, &self.test, &*self.err_match)
+    }
+}
+
+impl<T: FromStr> InputBuilderOnce<T> {
+    pub fn msg(self, msg: impl ToString) -> Self {
+        Self {
+            builder: self.builder.msg(msg),
+            ..self
+        }
+    }
+    pub fn repeat_msg(self, msg: impl ToString) -> Self {
+        Self {
+            builder: self.builder.repeat_msg(msg),
+            ..self
+        }
+    }
+    pub fn err(self, err: impl ToString) -> Self {
+        Self {
+            builder: self.builder.err(err),
+            ..self
+        }
+    }
+    pub fn add_test<F: 'static + Fn(&T) -> bool>(self, test: F) -> Self {
+        Self {
+            builder: self.builder.add_test(test),
+            ..self
+        }
+    }
+    pub fn add_err_test<F: 'static + Fn(&T) -> bool>(self, test: F, err: impl ToString) -> Self {
+        Self {
+            builder: self.builder.add_err_test(test, err),
+            ..self
+        }
+    }
+    pub fn clear_tests(self) -> Self {
+        Self {
+            builder: self.builder.clear_tests(),
+            ..self
+        }
+    }
+    pub fn err_match<F: 'static + Fn(&T::Err) -> Option<String>>(self, err_match: F) -> Self {
+        Self {
+            builder: self.builder.err_match(err_match),
+            ..self
+        }
+    }
     pub fn get(self) -> T {
         read_input::<T>(
-            &self.msg,
-            self.err,
+            &self.builder.msg,
+            &self.builder.err,
             self.default,
-            &self.test,
-            &*self.err_match,
+            &self.builder.test,
+            &*self.builder.err_match,
         )
     }
 }
 
-pub fn with_description<T: Error>(x: &T) -> Option<String> {
-    Some(format!("Error \"{}\"", (*x).description()))
-}
-
 /// Creates a new instance of `InputBuilder` with default settings. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-pub fn input_new<'a, T: FromStr>() -> InputBuilder<'a, T> {
+pub fn input_new<T: FromStr>() -> InputBuilder<T> {
     InputBuilder::new()
 }
 
-/// Shortcut function. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-pub fn valid_input<T: FromStr>(test: impl Fn(&T) -> bool + 'static) -> T {
-    input_new::<T>().add_test(test).get()
+fn try_flush() {
+    io::stdout().flush().unwrap_or(())
 }
 
-/// Shortcut function. This is documented in the [readme](https://gitlab.com/efunb/read_input/blob/master/README.md)
-pub fn simple_input<T: FromStr>() -> T {
-    input_new().get()
-}
-
-fn read_input<'a, T: FromStr>(
-    prompt: &PromptMsg<'a>,
-    err: &str,
-    default: Option<T>,
-    test: &[(Box<TestFunc<T>>, Option<&'a str>)],
-    err_pass: &dyn Fn(&T::Err) -> Option<String>,
-) -> T {
-    print!("{}", prompt.msg);
-    io::stdout().flush().expect("could not flush output");
-
+fn input_str() -> String {
     let mut input = String::new();
     io::stdin()
         .read_line(&mut input)
         .expect("Failed to read line");
+    input
+}
+
+fn read_input<T: FromStr>(
+    prompt: &PromptMsg,
+    err: &str,
+    default: Option<T>,
+    test: &[(Box<TestFunc<T>>, Option<String>)],
+    err_pass: &dyn Fn(&T::Err) -> Option<String>,
+) -> T {
+    print!("{}", prompt.msg);
+    try_flush();
+
+    let mut input = input_str();
 
     if input.trim().is_empty() {
         if let Some(x) = default {
             return x;
         }
     };
+
     loop {
         match T::from_str(&input.trim()) {
             Ok(value) => {
@@ -169,29 +229,26 @@ fn read_input<'a, T: FromStr>(
                     if f.0(&value) {
                         true
                     } else {
-                        test_err = Some(f.1.unwrap_or(err));
+                        test_err = Some(f.1.clone().unwrap_or_else(|| err.to_string()));
                         false
                     }
                 });
                 if passes_test {
                     return value;
                 } else {
-                    println!("{}", test_err.unwrap_or(err));
+                    println!("{}", test_err.unwrap_or_else(|| err.to_string()));
                 }
             }
             Err(error) => {
-                println!("{}", err_pass(&error).unwrap_or_else(|| err.to_string()));
+                println!("{}", (err_pass(&error)).unwrap_or_else(|| err.to_string()));
             }
         }
 
         if prompt.repeat {
             print!("{}", prompt.msg);
-            io::stdout().flush().expect("could not flush output");
+            try_flush();
         };
 
-        input.clear();
-        io::stdin()
-            .read_line(&mut input)
-            .expect("Failed to read line");
+        input = input_str();
     }
 }