+++ src/prelude.rs
@@ -1,0 +1,2 @@
+//TODO only put `shortcut` into scope without functions on inside.
+pub use crate::input_new;